Changing int to float in the point structure

+++ include/point.h
@@ -13,8 +13,8 @@
 
 /**
  * Creating a new point type based on a struct containing:
- * int x: horizontal coordinate of the point
- * int y: vertical coordinate of the point
+ * float x: horizontal coordinate of the point
+ * float y: vertical coordinate of the point
 */
 typedef struct point Point;
 
@@ -25,7 +25,7 @@ typedef struct point Point;
  * Conditions: none
  * Side effects: the structure is DYNAMICALLY allocated, remember to destroy it later!!!
 */
-Point *Point_Create(int x, int y);
+Point *Point_Create(float x, float y);
 
 /**
  * Declaring the function that destroys a given point:
@@ -79,7 +79,7 @@ bool Point_IsMarked(Point* point);
  * Conditions: existent and allocated point
  * Side effects: none
 */
-int Point_GetX(Point *point);
+float Point_GetX(Point *point);
 
 /**
  * Declaring the function that gets the y coordinate of the point:
@@ -88,6 +88,6 @@ int Point_GetX(Point *point);
  * Conditions: existent and allocated point
  * Side effects: none
 */
-int Point_GetY(Point *point);
+float Point_GetY(Point *point);
 
 #endif // POINT_H_
